+++ src/Adapter/Propel2Adapter.php
@@ -13,10 +13,13 @@ namespace Pagerfanta\Adapter;
 
 use Propel\Runtime\ActiveQuery\ModelCriteria;
 
+trigger_deprecation('pagerfanta/pagerfanta', '2.2', 'The "%s" adapter is deprecated and will be removed in 3.0.', Propel2Adapter::class);
+
 /**
  * Propel2Adapter.
  *
  * @author Raphael YAN <raphael.yan89@gmail.com>
+ * @deprecated to be removed in 3.0
  */
 class Propel2Adapter implements AdapterInterface
 {