+++ src/Adapter/PropelAdapter.php
@@ -11,10 +11,13 @@
 
 namespace Pagerfanta\Adapter;
 
+trigger_deprecation('pagerfanta/pagerfanta', '2.2', 'The "%s" adapter is deprecated and will be removed in 3.0.', PropelAdapter::class);
+
 /**
  * PropelAdapter.
  *
  * @author William DURAND <william.durand1@gmail.com>
+ * @deprecated to be removed in 3.0
  */
 class PropelAdapter implements AdapterInterface
 {